+++ src/main/java/dk/statsbiblioteket/newspaper/promptdomsingester/component/ConfigConstants.java
@@ -1,0 +1,12 @@
+package dk.statsbiblioteket.newspaper.promptdomsingester.component;
+
+/**
+ * Config constants specific to this component.
+ */
+public class ConfigConstants {
+
+    private ConfigConstants() {}
+
+    public static final String MAX_DOMS_UPDATE_RETRIES = "doms.update.retries";
+
+}