+++ src/main/java/dk/statsbiblioteket/newspaper/promptdomsingester/component/RunnableMultiThreadedPromptDomsIngester.java
@@ -48,7 +48,8 @@ public class RunnableMultiThreadedPromptDomsIngester extends TreeProcessorAbstra
         MultiThreadedFedoraIngester ingester = new MultiThreadedFedoraIngester(
                 eFedora,
                 new String[]{getProperties().getProperty(ConfigConstants.DOMS_COLLECTION, "doms:Newspaper_Collection")},
-                parseInt(getProperties().getProperty(ConfigConstants.THREADS_PER_BATCH))
+                parseInt(getProperties().getProperty(ConfigConstants.THREADS_PER_BATCH)),
+                parseInt(getProperties().getProperty(dk.statsbiblioteket.newspaper.promptdomsingester.component.ConfigConstants.MAX_DOMS_UPDATE_RETRIES))
         );
         try {
             ingester.ingest(iterator);