+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 1.4
 * Fixed https://sbprojects.statsbiblioteket.dk/jira/browse/NO-210 - duplicate relations created on repeated ingest.
+* Added new config parameter doms.update.retries (recommended value 2)
 
 1.3
 * Updated to newspaper-parent 1.2